requirements.txt
ipdb==0.7
ipython==0.13.2
ply==3.4
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2
